fixed gateway jwt filter, deleted jwtf in services

+++ barseek-bar-mservice/src/main/java/com/example/barseek_bar_mservice/controller/BarController.java
@@ -2,6 +2,7 @@ package com.example.barseek_bar_mservice.controller;
 
 
 import com.example.barseek_bar_mservice.dto.UpdateBarRequest;
+import com.example.barseek_bar_mservice.exception.customExceptions.UnauthorizedAccessException;
 import com.example.barseek_bar_mservice.model.entity.Bar;
 import com.example.barseek_bar_mservice.model.entity.Drink;
 import com.example.barseek_bar_mservice.security.UserPrincipal;
@@ -23,9 +24,13 @@ public class BarController {
 
     @PostMapping
     public ResponseEntity<String> addNew(@RequestBody Bar bar,
-                                         @AuthenticationPrincipal UserPrincipal principal
+                                         @RequestHeader(value = "X-User-Id") String userId,
+                                         @RequestHeader(value = "X-User-Roles") String roles
     ) {
-            Long ownerId = Long.parseLong(principal.getUsername());
+
+            if(!(roles.contains("bar"))) throw new UnauthorizedAccessException("User is not a bar owner!");
+
+            Long ownerId = Long.parseLong(userId);
             Bar newBar = barService.addNewBar(bar,ownerId);
             return new ResponseEntity<>("New bar created with name : " + newBar.getName(),HttpStatus.CREATED);
     }
@@ -48,9 +53,13 @@ public class BarController {
 
     @DeleteMapping("/{barId}")
     public ResponseEntity<String> deleteById(@PathVariable("barId") Long id,
-                                             @AuthenticationPrincipal UserPrincipal principal
+                                             @RequestHeader(value = "X-User-Id") String userId,
+                                             @RequestHeader(value = "X-User-Roles") String roles
+
     ) {
-            Long ownerId = Long.parseLong(principal.getUsername());
+            if(!(roles.contains("bar"))) throw new UnauthorizedAccessException("User is not a bar owner!");
+
+            Long ownerId = Long.parseLong(userId);
             barService.deleteBarById(id,ownerId);
             return ResponseEntity.ok(("Bar was deleted, id : " + id));
     }
@@ -58,9 +67,14 @@ public class BarController {
     @PutMapping("/{barId}")
     public ResponseEntity<String> updateById(@PathVariable("barId") Long id,
                                              @RequestBody UpdateBarRequest updateBarRequest,
-                                             @AuthenticationPrincipal UserPrincipal principal
+                                             @RequestHeader(value = "X-User-Id") String userId,
+                                             @RequestHeader(value = "X-User-Roles") String roles
+
     ) {
-            Long ownerId = Long.parseLong(principal.getUsername());
+
+            if(!(roles.contains("bar"))) throw new UnauthorizedAccessException("User is not a bar owner!");
+
+            Long ownerId = Long.parseLong(userId);
             Bar newBar = barService.updateBarById(id, updateBarRequest, ownerId);
             return new ResponseEntity<>("Bar was updated and saved, name : " + newBar.getName(), HttpStatus.OK);
     }
